Reduce spawn pool size
The vanilla spawn pool size is so large that it is pretty much impossible
to exhaust it before the assault times out even though ending the assault
early through killing a lot is an actual feature. The new numbers will
still make it hard to do so but there's an actual chance now.