Adjust damage network messages for health granularity
The new health granularity requires network messages that send damage to accept higher max values.
Health granularity in vanilla is 512 steps, while the new granularity in this mod is 8192.
While this granularity was already in effect, the network messages have not been adjusted until now which would lead to damage desync between clients and the host.