Reduce damage of some weapon presets slightly
Revolver and pistol preset were set slightly too high, not accounting for the extra multiplier that most of the units using these weapons get.
Rifle/SMG damage was a bit too high considering the minimum burst size is 3 shots as opposed to vanilla where enemies are way more likely to 1 or 2 tap.