Make enemies turn less sluggish when facing fast moving players
Enemies decide to turn when their forward direction differs too much from the current player position.
If the player is close to them and moves fast, the position is outdated by the time the enemy has turned.
This change will add the player's velocity to the requested turn position as a kind of position prediction to avoid this issue.