Make enemy flank pathing more dynamic
Enemies that try to avoid/flank players do so by checking for unsafe navigation segments. By default, navigation segments are only marked as unsafe if a player is there.
The change in this commit marks a nav segment unsafe for a short duration whenever an enemy dies in it, which should make enemy pathing slighty more dynamic for groups that use the safe check.