Make shields block explosion damage from the front
Shield units that are in between the center of an explosion and other units (including themselves) will now reduce the explosion damage those units would receive.
This change was made in favor over the generic explosion resistance shields had before and now only applies damage reduction if the shield would actually block the explosion.
If an explosion occurs behind or to the side of the shield the full damage still applies.