Fix enemy shooting actions syncing incorrectly to clients
Whenever an enemy starts/stops shooting on the host, it sends a message to clients to do the same, but the parameter to start/stop shooting is ignored and there's always a shooting action started on the client, even if it should be stopped.
This issue is what clients would most likely experience as enemies shooting at them while they are doing special navigation animations like climbing.