Improve NPC accuracy calculation
Now adds a minimum amount of spread that is purely cosmetic if the player is completely unobstructed.
If the player is partly covered by other objects, such as being in a vehicle, that spread can lead to NPCs hitting the cover instead.
This should improve the experience in vehicles, which previously offered no extra cover when enemies had direct line of sight due to them aiming at the exact center of the player camera.